use absolute positioning for chapters
